dbesmpm.py
This is the code used to create the database schema and populate it with initial data. It defines the structure of the database tables and inserts predefined values. This creation code is not required for normal usage.

Future updates to the database will involve adding new models and possibly modifying the schema to accommodate additional types of data relevant to protein evolution research